backend/OLDrequirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
coverage==5.2.1
cryptography==2.9.2
daphne==2.5.0
Django==3.1.2
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
# Khayyam==3.0.17
kombu==4.6.11
mccabe==0.6.1
msgpack==1.0.0
# numpy==1.19.0
# pandas==1.0.5
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
